+++ dash_bio/component_factory/_clustergram.py
@@ -2,18 +2,19 @@
 from random import shuffle
 
 import numpy as np
+import pandas as pd
 import scipy
 import scipy.cluster.hierarchy as sch
 import scipy.spatial as scs
 from sklearn.impute import SimpleImputer
 
-import plotly.graph_objs as go
+import plotly.graph_objects as go
 from plotly import tools
 
 
 # pylint: disable=assignment-from-no-return, no-self-use
 def Clustergram(
-        data=None,
+        data,
         generate_curves_dict=False,
         return_computed_traces=False,
         computed_traces=None,
@@ -49,7 +50,8 @@ def Clustergram(
 
 Keyword arguments:
 
-- data (ndarray; required): Matrix of observations as array of arrays
+- data (2D array-like; required): Matrix or table of observations (dropping
+    columns of non-numeric dtype).
 - generate_curves_dict (bool; default False): Whether or not to return a
     dictionary containing information about the cluster number
     associated with each curve number in the graph. (May be useful
@@ -64,8 +66,8 @@ Keyword arguments:
    (observation labels).
 - column_labels (list; optional): List of column category labels
    (observation labels).
-- hidden_labels (list; optional): List of labels not to display on the
-    final plot.
+- hidden_labels (list; optional): List containing strings 'row' and/or 'col'
+    if row and/or column labels should be hidden on the final plot.
 - standardize (string; default 'none'): The dimension for standardizing
     values, so that the mean is 0 and the standard deviation is 1,
     along the specified dimension: 'row', 'column', or 'none'.
@@ -143,11 +145,11 @@ Keyword arguments:
     'color' (a string representation of the color of the
     annotation).
 - tick_font (dict; optional): The font options for ticks, as specified
-    in the Plotly graph_objs documentation (see:
+    in the Plotly graph_objects documentation (see:
     https://plotly.com/python/reference/#bar-marker-colorbar-tickfont).
 - annotation_font (dict; optional): The font options for annotations,
-    as specified in the Plotly graph_objs documentation (see:
-    https://plotly.com/python/reference/#layout-scene-annotations-items-annotation-font).
+    as specified in the Plotly graph_objects documentation (see:
+    https://plotly.cp,/python/reference/#layout-scene-annotations-items-annotation-font).
 - line_width (list | number; default 0.5): The line width for the
     dendrograms. If in list format, the first element corresponds to
     the width of the row dendrogram traces, and the second corresponds
@@ -160,8 +162,6 @@ Keyword arguments:
 - width (number; default 500): The width of the graph, in px.
 
     """
-    if hidden_labels is None:
-        hidden_labels = []
     if color_threshold is None:
         color_threshold = dict(row=0, col=0)
 
@@ -195,12 +195,12 @@ class _Clustergram:
 
 Methods:
 
-- figure(computed_traces=None): Return a figure object compatible with plotly.graph_objs.
+- figure(computed_traces=None): Return a figure object compatible with plotly.graph_objects.
     """
 
     def __init__(
             self,
-            data=None,
+            data,
             row_labels=None,
             column_labels=None,
             hidden_labels=None,
@@ -234,20 +234,23 @@ Methods:
     See docstring of the `Clustergram` function, where the same keyword arguments (and a couple
     of other ones) are documented.
         """
+        if isinstance(data, pd.DataFrame):
+            data = data.select_dtypes('number')
+            data = data.values
         if hidden_labels is None:
             hidden_labels = []
         if color_threshold is None:
             color_threshold = dict(row=0, col=0)
-        if row_labels is None:
-            row_labels = [str(i) for i in range(data.shape[0])]
-            hidden_labels.append("row")
-        if column_labels is None:
-            column_labels = [str(i) for i in range(data.shape[1])]
-            hidden_labels.append("col")
+        # Always keep unique identifiers for rows
+        row_ids = list(range(data.shape[0]))
+        # Always keep unique identifiers for columns
+        column_ids = list(range(data.shape[1]))
 
         self._data = data
         self._row_labels = row_labels
+        self._row_ids = row_ids
         self._column_labels = column_labels
+        self._column_ids = column_ids
         self._cluster = cluster
         self._row_dist = row_dist
         self._col_dist = col_dist
@@ -341,7 +344,7 @@ Methods:
             self._data = self._scale(standardize)
 
     def figure(self, computed_traces=None):
-        """Return a figure object compatible with plotly.graph_objs.
+        """Return a figure object compatible with plotly.graph_objects.
 
     Parameters:
 
@@ -354,16 +357,23 @@ Methods:
             (
                 dt,
                 self._data,
-                self._row_labels,
-                self._column_labels,
+                self._row_ids,
+                self._column_ids,
             ) = self._compute_clustered_data()
         else:
             # use, if available, the precomputed dendrogram and heatmap
             # traces (as well as the row and column labels)
             dt = computed_traces["dendro_traces"]
             heatmap = computed_traces["heatmap"]
-            self._row_labels = computed_traces["row_labels"]
-            self._column_labels = computed_traces["column_labels"]
+            self._row_ids = computed_traces["row_ids"]
+            self._column_ids = computed_traces["column_ids"]
+
+        # Match reordered rows and columns with their respective labels
+        if self._row_labels:
+            self._row_labels = [self._row_labels[r] for r in self._row_ids]
+        if self._column_labels:
+            self._column_labels = [self._column_labels[r]
+                                   for r in self._column_ids]
 
         # this dictionary relates curve numbers (accessible from the
         # hoverData/clickData props) to cluster numbers
@@ -496,7 +506,7 @@ Methods:
         xaxis2.update(scaleanchor="x5")
 
         if len(tickvals_col) == 0:
-            tickvals_col = [10 * i + 5 for i in range(len(self._column_labels))]
+            tickvals_col = [10 * i + 5 for i in range(len(self._column_ids))]
 
         # add in all of the labels
         fig["layout"]["xaxis5"].update(  # pylint: disable=invalid-sequence-index
@@ -513,7 +523,7 @@ Methods:
         )
 
         if len(tickvals_row) == 0:
-            tickvals_row = [10 * i + 5 for i in range(len(self._row_labels))]
+            tickvals_row = [10 * i + 5 for i in range(len(self._row_ids))]
 
         fig["layout"]["yaxis5"].update(  # pylint: disable=invalid-sequence-index
             tickmode="array",
@@ -671,8 +681,8 @@ Methods:
         computed_traces = {
             "dendro_traces": dt,
             "heatmap": heatmap,
-            "row_labels": self._row_labels,
-            "column_labels": self._column_labels,
+            "row_ids": self._row_ids,
+            "column_ids": self._column_ids,
         }
 
         return (fig, computed_traces, cluster_curve_numbers)
@@ -741,8 +751,8 @@ Methods:
         # first, compute the clusters
         (Zcol, Zrow) = self._get_clusters()
 
-        clustered_column_labels = self._column_labels
-        clustered_row_labels = self._row_labels
+        clustered_column_ids = self._column_ids
+        clustered_row_ids = self._row_ids
 
         # calculate dendrogram from clusters; sch.dendrogram returns sets
         # of four coordinates that make up the 'u' shapes in the dendrogram
@@ -751,10 +761,10 @@ Methods:
                 Zcol,
                 orientation="top",
                 color_threshold=self._color_threshold["col"],
-                labels=self._column_labels,
+                labels=self._column_ids,
                 no_plot=True,
             )
-            clustered_column_labels = Pcol["ivl"]
+            clustered_column_ids = Pcol["ivl"]
             trace_list["col"] = self._color_dendro_clusters(Pcol, "col")
 
         if Zrow is not None:
@@ -762,7 +772,7 @@ Methods:
                 Zrow,
                 orientation="left",
                 color_threshold=self._color_threshold["row"],
-                labels=self._row_labels,
+                labels=self._row_ids,
                 no_plot=True,
             )
             # need to flip the coordinates for the row dendrogram
@@ -771,21 +781,21 @@ Methods:
                 "dcoord": Prow["icoord"],
                 "color_list": Prow["color_list"],
             }
-            clustered_row_labels = Prow["ivl"]
+            clustered_row_ids = Prow["ivl"]
             trace_list["row"] = self._color_dendro_clusters(Prow_tmp, "row")
 
         # now, we need to rearrange the data array to fit the labels
 
         # first get reordered indices
-        rl_indices = [self._row_labels.index(r) for r in clustered_row_labels]
-        cl_indices = [self._column_labels.index(c) for c in clustered_column_labels]
+        rl_indices = [self._row_ids.index(r) for r in clustered_row_ids]
+        cl_indices = [self._column_ids.index(c) for c in clustered_column_ids]
 
         # modify the data here; first shuffle rows,
         # then transpose and shuffle columns,
         # then transpose again
         clustered_data = self._data[rl_indices].T[cl_indices].T
 
-        return trace_list, clustered_data, clustered_row_labels, clustered_column_labels
+        return trace_list, clustered_data, clustered_row_ids, clustered_column_ids
 
     def _color_dendro_clusters(self, P, dim):
         """Color each cluster below the color threshold separately.