+++ tests/integration/test_clustergram.py
@@ -1,16 +1,16 @@
-import pandas
 import json
+import pandas as pd
 
 import dash
-import dash_bio
 import dash_html_components as html
+import dash_bio
 
 from common_features import nested_component_layout, \
     nested_component_app_callback
 
 _data = None
 
-_mtcars_data = pandas.read_csv(
+_mtcars_data = pd.read_csv(
     'tests/dashbio_demos/dash-clustergram/data/mtcars.tsv',
     delimiter='\t',
     skiprows=4
@@ -185,3 +185,91 @@ def test_dbcl005_row_annotations(dash_duo):
     # the annotation is the correct color
     dash_duo.wait_for_style_to_equal(
         'g.subplot.x6y6 g.plot g.lines > path', 'stroke', 'rgb(248, 62, 199)')
+
+
+def test_dbcl006_df_input_row_cluster(dash_duo):
+
+    app = dash.Dash(__name__)
+
+    # run the same test as dbcl002 (row clustering) where table of
+    # observations (data argument) is left as a DataFrame
+    assert isinstance(_mtcars_data, pd.DataFrame)
+    app.layout = html.Div(nested_component_layout(
+        dash_bio.Clustergram(
+            data=_mtcars_data
+        )
+    ))
+
+    nested_component_app_callback(
+        app,
+        dash_duo,
+        component=dash_bio.Clustergram,
+        component_data=_data,
+        test_prop_name='cluster',
+        test_prop_value='row',
+        prop_value_type='string'
+    )
+
+    assert len(dash_duo.find_elements('g.subplot.x2y2')) == 0
+    assert len(dash_duo.find_elements('g.subplot.x4y4')) == 1
+
+
+def test_dbcl007_hidden_labels(dash_duo):
+
+    app = dash.Dash(__name__)
+
+    data = _mtcars_data
+    row_labels = list(_mtcars_data.index)
+    col_labels = list(_mtcars_data.columns)
+
+    app.layout = html.Div(nested_component_layout(
+        dash_bio.Clustergram(
+            data=data,
+            row_labels=row_labels,
+            column_labels=col_labels
+        )
+    ))
+
+    nested_component_app_callback(
+        app,
+        dash_duo,
+        component=dash_bio.Clustergram,
+        component_data=data,
+        test_prop_name='hidden_labels',
+        test_prop_value='row',
+        prop_value_type='string'
+    )
+
+    # ensure that row labels are hidden
+    assert len(dash_duo.find_elements('g.yaxislayer-above g.y5tick')) == 0
+    # ensure that column labels are displayed
+    assert len(dash_duo.find_elements('g.xaxislayer-above g.x5tick')) == \
+        len(col_labels)
+
+    # create a new instance of the app to test hiding of column labels
+
+    app = dash.Dash(__name__)
+
+    app.layout = html.Div(nested_component_layout(
+        dash_bio.Clustergram(
+            data=data,
+            row_labels=row_labels,
+            column_labels=col_labels
+        )
+    ))
+
+    nested_component_app_callback(
+        app,
+        dash_duo,
+        component=dash_bio.Clustergram,
+        component_data=data,
+        test_prop_name='hidden_labels',
+        test_prop_value='col',
+        prop_value_type='string'
+    )
+
+    # ensure that column labels are hidden
+    assert len(dash_duo.find_elements('g.xaxislayer-above g.x5tick')) == 0
+    # ensure that row labels are displayed
+    assert len(dash_duo.find_elements('g.yaxislayer-above g.y5tick')) == \
+        len(row_labels)