+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [0.4.8] - 2020-03-16
+### Changed
+* [#489](https://github.com/plotly/dash-bio/pull/489) Renamed async modules with hyphen `-` instead of tilde `~`
+* [#494](https://github.com/plotly/dash-bio/pull/494) Update from React 16.8.6 to 16.13.0
+
+### Added
+* [#492](https://github.com/plotly/dash-bio/pull/492) Added working support of
+  labels in Clustergram.
+
+## [0.4.7] - 2020-02-21
+### Added
+* [#478](https://github.com/plotly/dash-bio/pull/478) Added support of
+  DataFrame as Clustergram input.
+
 ## [0.4.6] - 2020-01-07
 ### Fixed
 * [#458](https://github.com/plotly/dash-bio/pull/458) Fixed reordering bug of